+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation_objc/include/camera_avfoundation/FLTCam.h
@@ -42,10 +42,14 @@ NS_ASSUME_NONNULL_BEGIN
 @property(assign, nonatomic) BOOL isRecording;
 @property(assign, nonatomic) BOOL isRecordingPaused;
 @property(strong, nonatomic, nullable) NSObject<FLTAssetWriter> *videoWriter;
-@property(assign, nonatomic) BOOL videoIsDisconnected;
-@property(assign, nonatomic) BOOL audioIsDisconnected;
-@property(assign, nonatomic) CMTime videoTimeOffset;
-@property(assign, nonatomic) CMTime audioTimeOffset;
+/// Whether the recording is disconnected.
+@property(assign, nonatomic) BOOL isRecordingDisconnected;
+/// Represents sum of all recording pauses/interruptions.
+@property(assign, nonatomic) CMTime recordingTimeOffset;
+/// Output to use for adjusting of recording offset.
+@property(nonatomic) AVCaptureOutput *outputForOffsetAdjusting;
+/// Time of the last appended video sample.
+@property(assign, nonatomic) CMTime lastAppendedVideoSampleTime;
 @property(strong, nonatomic, nullable) NSObject<FLTAssetWriterInput> *videoWriterInput;
 @property(strong, nonatomic, nullable) NSObject<FLTAssetWriterInput> *audioWriterInput;
 @property(nullable) NSObject<FLTAssetWriterInputPixelBufferAdaptor> *videoAdaptor;